[REFACTO] rework of Issue's Modals ( HTML + CSS )
