+++ front/components/Modals/issue/deleteIssueModal.vue
@@ -45,11 +45,14 @@ onMounted(() => {
     <div class="modal-background"></div>
     <div class="modal-content">
       <div class="box del-box">
-        <h3>Are you sure you want to delete the issue {{ project?.name.substring(0, 3).toUpperCase() }}-{{ issue?.ticket_id }}?</h3>
-        <div class="btn-action">
-          <button class="cancel" @click="showModal=false">Cancel</button>
-          <button class="del" @click="handleDelete()">Delete</button>
-        </div>
+        <h4>Are you sure you want to delete the issue {{ project?.name.substring(0, 3).toUpperCase() }}-{{ issue?.ticket_id }}?</h4>
+        <!--       Btn action       -->
+        <button class="button is-cancel" @click="showModal = false"> <!-- blanc + border gris -->
+          Cancel
+        </button>
+        <button class="button is-danger" @click="showModal=false; handleDelete()">
+          Delete
+        </button>
       </div>
     </div>
   </div>
